[code quality] documentation and cleanup
