+++ src/SDDK/GPU/acc_runtime.hpp
@@ -17,9 +17,9 @@
 // CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
 // OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-/** \file acc.hpp
+/** \file acc_runtime.hpp
  *
- *  \brief Interface to accelerators API.
+ *  \brief Uniform interface to the runtime API of CUDA and ROCm.
  *
  */
 #ifndef __ACC_RUNTIME_HPP__
